Handler, message_routers.go: added CRUD actions.

+++ internal/models/resource_model.go
@@ -5,7 +5,7 @@ import "time"
 type Resource struct {
 	ID          int    `copier:"-"`
 	UserId      int    `copier:"-"`
-	Code        int    `copier:"-"`
+	Code        int64  `copier:"-"`
 	URL         string `copier:"URL"`
 	Description string `copier:"Description"`
 	CreatedAt   time.Time
